[SDFAB-158] eNB not able to get IP through fabric
After subsequent calls to configureNetwork() we dont set
the onComplete handler and due to this the queue events
are never processed. Most of the times everything works
despite this evident bug because in the DeviceConfiguration
we parse in a greedy way all the netcfg. Sometimes this can
fail if all the netcfg is not immediately available.

Likely fix [SDFAB-113] SR sometimes does not create flows and groups

Change-Id: Iebfa9bf9b1cc29af112bf38895441cbac47c7774