[SDFAB-359] Purge all on device disconnection
The previous logic was relying on the purgeOnDisconnection device configuration.
However, we want to allow flow rules to stay in the store when a device disconnects if those are
created by a different application. For this reason, we purge flow rule and groups explicitly on
device disconnection in SR by specifying the SR app ID.

Change-Id: I59369a1a18b82f134aaaab37be49e2ac746a4e9f