Revert "SDFAB-242 Explicitly use wipeDeferred in the flow obj to drop ARP and NDP from pair port"
This reverts commit 90dd932cd30d5032a789120baeee96d245b47005.

Reason for revert: With this patch we are no more able to forward ARP/NDP traffic through the bridging using the pair link. Going back in the past - it looks like the original purpose was to prevent learning of hosts through the pair link and not explicitly dropped the ARP requests.

Change-Id: I8e0547c444b471c364925fd3eb1cb8197789ede6